Exp1_DEMO.py: change run_number name in drop down menu so it appears 2nd in list.  Changed directions for target_jump in top-right (45) or bottom left (225) as they were back-to-front. Exp1_direction_detection_2AFC.py: Task to see if participants can detect motion in probes: 2AFC, did you see clockwise or anti-clockwise motion. Exp1_R_G.py: made probe_color255 into an int. Exp1_sep4_5.py: made probe_color255 into an int. Exp1_speed_detection.py: made probe_color255 into an int.

+++ Nick_scripts/Exp1_R_G.py
@@ -379,7 +379,7 @@ for step in range(n_trials_per_stair):
         target_jump = random.choice([1, -1])
         # staircase varied probeLum
         probeLum = thisStair.next()
-        probeColor255 = probeLum * LumColor255Factor
+        probeColor255 = int(probeLum * LumColor255Factor)
         probeColor1 = (probeColor255 * Color255Color1Factor) - 1
 
         print(f"probeLum: {probeLum}")
